# docs/swagger.yaml
definitions:
...
  dtousers.TeamIndexInfo:
    properties:
      avatar:
        description: 头像
        type: string
      children:
        description: 下级用户列表
        items:
          $ref: '#/definitions/dtousers.TeamIndexInfo'
        type: array
      createdAt:
        description: 注册时间
        type: integer
      depth:
        description: 深度
        type: integer
      earnings:
        description: 团队收益
        type: number
      id:
        description: 用户ID
        type: integer
      nickname:
        description: 用户昵称
        type: string
      username:
        description: 用户名
        type: string
    type: object
...
  dtowallets.HomeAccountInfo:
    properties:
      code:
        description: 银行代码
        type: string
      createdAt:
        description: 创建时间
        type: integer
      icon:
        description: 图标
        type: string
      id:
        description: 主键
        type: integer
      name:
        description: 类型名称
        type: string
      number:
        description: 卡号
        type: string
      paymentId:
        description: 支付ID
        type: integer
      paymentName:
        description: 支付名称
        type: string
      paymentType:
        description: 支付类型
        type: integer
      realName:
        description: 真实姓名
        type: string
      status:
        description: 状态
        type: integer
    type: object
...
  dtowallets.HomeWalletOrderIndexData:
    properties:
      data:
        description: 拒绝理由
        type: string
      fee:
        description: 手续费
        type: number
      icon:
        description: 图标
        type: string
      id:
        description: 主键
        type: integer
      money:
        description: 金额
        type: number
      name:
        description: 名称
        type: string
      status:
        description: 状态
        type: integer
      type:
        description: 类型 1充值类型 2资产充值类型 11提现类型 12资产提现类型
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  models.WalletBillTypeOptionsInfo:
    properties:
      label:
        description: 账单标题
        type: string
      value:
        description: 账单值
        type: integer
    type: object
...
  utils.Pagination:
    properties:
      descending:
        description: 排序 真DESC 假ASC
        type: boolean
      page:
        description: 当前页数
        type: integer
      rowsPerPage:
        description: 每页显示条数
        type: integer
      sortBy:
        description: 排序字段
        type: string
    type: object
...
  utils.ResponseJson:
    properties:
      code:
        description: 错误代码
        type: integer
      data:
        description: 内容
      msg:
        description: 错误信息
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/user/team/index:
    post:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 团队成员列表
          schema:
            $ref: '#/definitions/dtousers.TeamIndexInfo'
      summary: 获取团队成员列表
      tags:
      - User
...
  /api/v1/wallets/account/info:
    post:
      parameters:
      - description: 主键
        in: body
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResponseJson'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtowallets.HomeAccountInfo'
                  type: array
              type: object
      summary: 卡片详情
      tags:
      - Wallet
...
  /api/v1/wallets/bill/index:
    post:
      parameters:
      - description: 类型 1充值类型 11提现类型 21购买 51收益 61奖励
        in: body
        name: type
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResponseJson'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.WalletBillTypeOptionsInfo'
                  type: array
              type: object
      summary: 用户账单类型
      tags:
      - Wallet
...
  /api/v1/wallets/order/index:
    post:
      parameters:
      - description: 资产ID
        in: body
        name: assetsId
        schema:
          type: integer
      - description: 类型 1充值类型 2资产充值类型 11提现类型 12资产提现类型
        in: body
        name: types
        schema:
          items:
            type: integer
          type: array
      - description: 分页参数
        in: body
        name: pagination
        schema:
          $ref: '#/definitions/utils.Pagination'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResponseJson'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dtowallets.HomeWalletOrderIndexData'
                  type: array
              type: object
      summary: 钱包订单列表
      tags:
      - Wallet
...
swagger: "2.0"